refactor: use registry pattern for animation system
Each animation file now self-registers its component and resolver via
a shared registry. Adding a new animation only requires creating the
file and one import line in animationResolvers.js. Shared DOM helpers
extracted to domHelpers.js.
